src/frontend/requirements.txt
python-dotenv>=1.1.0
streamlit>=1.44.1,<2.0.0
